Use flow decomposition code from David Aulicino as a benchmark
David Aulicino shared a Jupyter notebook with us that (in June 2021)
took about 12h to complete. This adds a simplified version of that
notebook that contains the bits that take 99% of the runtime.

The computationally difficult part is computing all the flow
decompositions for directions coming from saddle connections up to some
length.

(In the original notebook that length was given as 400 for the square of
the actual length.)