Use Python warnings instead of SageMath deprecation warnings
the SageMath deprecation warnings point to trac pages for SageMath which
is confusing (and unrelated to the sage-flatsurf deprecation.)

I tried to use a package such as Deprecated or deprecation for this but
they don't play nicely with the doctest runner in SageMath. (The
doctests throw an exception because they do not know how to properly
instantiate a DeprecationWarning. Also, normal users won't see these
warnings since one has to set PYTHONWARNINGS=d which probably nobody
does. And we do really want non-developer users of sage-flatsurf to see
these warnings since there is no downstream developers for this project.)