Fix uniqueness of homology
we cannot have SimplicialHomology be a UniqueRepresentation since
surfaces are not unique representation. However, we still want
Homology(S) is Homology(S). We implement this by using a cached method
in the surface. There are some drawbacks with this approach, mostly with
pickling, we get Homology(S) is Homology(S) but Homology(loads(dumps(S))
is not Homology(S) and loads(dumps(Homology(S)) is not Homology(S).

Also there is the potential for a memory leak if homology with lots of
different coefficient rings get created.