Fix moduli computation in example
Fixes #235. https://arxiv.org/pdf/1411.1827 Definition 3.3 defines the
moduli as h/c where c is the circumference so we want to follow that
definition.

Note that we use that correct definition in gl2r_orbit_closure.py

```py
  width = self.V2._isomorphic_vector_space.base_ring()(
      self.V2.base_ring()(component.width())
  )
  height = self.V2._isomorphic_vector_space.base_ring()(
      self.V2.base_ring()(
          component.vertical().project(component.circumferenceHolonomy())
      )
  )
  module_fractions.append((width, height))
```

What's confusing is that "height" is the circumference here and width is
what is called height in the article (in the article, the flow is to the
right, for us the flow is usually north.)